# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Comoon API
  description: Social commerce platform connecting leaders, entrepreneurs, and causes in Colombia.
  version: 1.0.0
  contact:
    name: Comoon Team
servers:
  - url: https://comoon.pages.dev/api
    description: Production
  - url: http://localhost:4321/api
    description: Local development
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
...
    Leader:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        bio:
          type: string
        location:
          type: string
        city:
          type: string
        department:
          type: string
        photo_url:
          type: string
        tags:
          type: array
          items:
            type: string
        is_verified:
          type: boolean
...
    CauseUpdate:
      type: object
      properties:
        id:
          type: integer
        leader_id:
          type: integer
        cause_id:
          type: integer
        title:
          type: string
        content:
          type: string
        photo_url:
          type: string
        photos:
          type: array
          items:
            type: string
          nullable: true
        update_type:
          type: string
          enum: [progress, milestone, gratitude, closing]
        is_closing:
          type: boolean
        created_at:
          type: string
          format: date-time
        leader:
          type: object
          properties:
            id:
              type: integer
            name:
              type: string
            photo_url:
              type: string
            city:
              type: string
            department:
              type: string
        cause:
          type: object
          properties:
            id:
              type: integer
            title:
              type: string
            status:
              type: string
    CreateCauseUpdateRequest:
      type: object
      required: [cause_id, title, content]
      properties:
        cause_id:
          type: integer
        title:
          type: string
        content:
          type: string
        photo_url:
          type: string
        photos:
          type: array
          items:
            type: string
        update_type:
          type: string
          enum: [progress, milestone, gratitude, closing]
        is_closing:
          type: boolean
...
paths:
...
  /leaders:
    get:
      summary: List leaders with optional filters
      security: []
      tags: [Leaders]
      parameters:
        - name: department
          in: query
          schema:
            type: string
        - name: tag
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
        - name: offset
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of leaders
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Leader"
  /leaders/{id}:
    get:
      summary: Get leader by ID
      security: []
      tags: [Leaders]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Leader details
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: "#/components/schemas/Leader"
        "404":
          description: Leader not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /cause-updates:
    get:
      summary: List cause updates with pagination and filters
      security: []
      tags: [Cause Updates]
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            maximum: 50
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
        - name: leader_id
          in: query
          schema:
            type: integer
        - name: cause_id
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of cause updates
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CauseUpdate"
    post:
      summary: Create a cause update (leaders only)
      tags: [Cause Updates]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCauseUpdateRequest"
      responses:
        "201":
          description: Cause update created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CauseUpdate"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Cause not found or not owned by leader
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /cause-updates/{id}:
    get:
      summary: Get a single cause update
      security: []
      tags: [Cause Updates]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Cause update details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CauseUpdate"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    put:
      summary: Update a cause update (owner leader or admin)
      tags: [Cause Updates]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                content:
                  type: string
                photo_url:
                  type: string
                photos:
                  type: array
                  items:
                    type: string
                update_type:
                  type: string
                  enum: [progress, milestone, gratitude, closing]
      responses:
        "200":
          description: Cause update modified
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CauseUpdate"
        "403":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      summary: Delete a cause update (owner leader or admin)
      tags: [Cause Updates]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "403":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"